requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── LLM & Agent Orchestration ───────────────────────────────────────────────
groq==0.11.0
langgraph==0.2.48
langchain-core>=0.3.36
langchain-groq==0.2.1
langsmith==0.1.143

# ── MCP ─────────────────────────────────────────────────────────────────────
fastmcp==0.4.1
mcp>=1.2.1
langchain-mcp-adapters==0.0.3

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Vector DB ────────────────────────────────────────────────────────────────
qdrant-client==1.12.1
langchain-qdrant==0.2.0

# ── Cache / Queue ────────────────────────────────────────────────────────────
redis==5.2.0
celery==5.4.0
upstash-redis==1.1.0

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2

# ── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers==3.3.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-multipart==0.0.12
aiofiles==24.1.0
tenacity==9.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.27.2

bcrypt==4.0.1
passlib[bcrypt]==1.7.4